+++ package.yaml
@@ -95,18 +95,19 @@ library:
     - Juvix.Core.Parameterisations.Unit
     - Juvix.Utility
     - Juvix.Utility.HashMap
-    - Juvix.Nets.Bohm
-    - Juvix.Nets.Combinators
-    - Juvix.Backends.Interface
-    - Juvix.Backends.Env
-    - Juvix.Backends.Graph
-    - Juvix.Backends.Maps
-    - Juvix.Bohm.Parser
-    - Juvix.Bohm.Translation
-    - Juvix.Bohm.Type
-    - Juvix.Bohm.Shared
-    - Juvix.Bohm.Default
-    - Juvix.Bohm
+    - Juvix.Interpreter
+    - Juvix.Interpreter.InteractionNet
+    - Juvix.Interpreter.InteractionNet.Nets.Bohm
+    - Juvix.Interpreter.InteractionNet.Nets.Combinators
+    - Juvix.Interpreter.InteractionNet.Backends.Interface
+    - Juvix.Interpreter.InteractionNet.Backends.Env
+    - Juvix.Interpreter.InteractionNet.Backends.Graph
+    - Juvix.Interpreter.InteractionNet.Backends.Maps
+    - Juvix.Interpreter.InteractionNet.Parser
+    - Juvix.Interpreter.InteractionNet.Translation
+    - Juvix.Interpreter.InteractionNet.Type
+    - Juvix.Interpreter.InteractionNet.Shared
+    - Juvix.Interpreter.InteractionNet.Default
     - Juvix.Library
     - Juvix.Visualize.Dot
     - Juvix.Visualize.Graph
@@ -121,7 +122,7 @@ library:
     - Juvix.LLVM.Codegen.Block
     - Juvix.LLVM.Codegen.Graph
   other-modules:
-    - Juvix.NodeInterface
+    - Juvix.Interpreter.InteractionNet.NodeInterface
     - Juvix.Utility.Helper
     - Juvix.Utility.PrettyPrint
     - Juvix.Core.Translate
